requirements.txt
django==2.2.5
gunicorn
django-heroku
django-crispy-forms==1.7.2
django-allauth
social-auth-app-django
Pillow
django-money
django-phone-field
django-phonenumber-field
phonenumbers
django-storages
boto3
django-mapbox-location-field==0.2.5
django-location-field
geopy
twilio
django-twilio
django-filter
django-bootstrap-form
django-widget-tweaks
django-starfield==1.0.post1